requirements.txt
anyio==3.6.1
autopep8==1.7.0
awscrt==0.13.8
boto3==1.24.49
botocore==1.27.49
click==8.1.3
fastapi==0.79.0
h11==0.13.0
idna==3.3
jmespath==1.0.1
pycodestyle==2.9.1
pydantic==1.9.1
python-dateutil==2.8.2
python-magic==0.4.27
python-multipart==0.0.5
s3transfer==0.6.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
